checkpoint

+++ P04/store/Plant.java
@@ -1,6 +1,6 @@
 package.store;
 public class Plant extends Product(){
-    public Plant(String species, double price, Exposure exposure){
+    public Plant(String species, int price, Exposure exposure){
         super(species, price);
         this.exposure = exposure;
     }
